+++ src/main/java/com/lucendar/gnss/sdk/evtsrc/GnssEventListener.java
@@ -1,0 +1,16 @@
+package com.lucendar.gnss.sdk.evtsrc;
+
+public interface GnssEventListener {
+
+    void afterSubscribed(GnssEventType eventType);
+
+    /**
+     *
+     * @param eventType
+     * @param t
+     * @return true if reconnect needed
+     */
+    void onFailure(GnssEventType eventType, Throwable t);
+
+    void onEvent(GnssEvent event);
+}