+++ src/main/java/com/lucendar/gnss/sdk/types/GnssEventListener.java
@@ -1,6 +1,0 @@
-package com.lucendar.gnss.sdk.types;
-
-public interface GnssEventListener {
-
-    void on(GnssEvent event);
-}