+++ src/main/java/com/lucendar/gnss/sdk/types/GnssEventType.java
@@ -1,8 +1,0 @@
-package com.lucendar.gnss.sdk.types;
-
-public enum GnssEventType {
-    cmd,
-    av_upload,
-    strm,
-    after_connect;
-}