+++ doc/change-log.md
@@ -1,3 +1,12 @@
+# 3.4.2
+## Improvement
+- Add android support
+- `GnssOpenStrmReq` add `inputAudioCfg` property
+
+## Changed
+- Use okhttp to implement websocket
+- Remove spring dependencies
+
 # 1.1.0
 ## Improvement
 - Introduce GnssApi, GnssWs